Add convenience for open_resource_type
"module_name" argument is, per documentation, not yet implemented and must always be null. With the added function signature in this commit it is defaulted to nil. Also, the resource destructor is also defaulted to nil. This simplified form of open_resource type only applies to cases where resource will be released with enif_release_resource().